refactor: collapse the two Sections into one, on the base
`Section`, `Toolbar`, `Actions` and `Facts` move out of `ui/controls.tsx`
into `ui/section.tsx`, rebuilt on Tailwind with their props unchanged.
Twenty-one consumers change one import line.

**There were two `Section`s.** One in the old control set and one in the
shell, drawn from the same three stylesheet classes, differing only in
whether the head was a `<header>` and whether the action was wrapped in a
`<div>`. Thirteen callers used the first and one — `run-status.tsx` — used
the second. Ticket 14 asked whether the product keeps one control vocabulary
or two; there is now one `Section` and the `<header>` form won it, because
that is what thirteen of the fourteen callers were already being given and
because it is what the element is.

`agents-pages.test.tsx` asserted the shared form layout by looking for
`form` and `formActions` in a class name. That worked while a CSS Module
hashed the only class that could contain those words; on the Tailwind base a
class list is copyable, so the check would have passed a hand-rolled `<div>`
— and it failed the migration, which is the same weakness from the other
side. It now asks the elements what they are, through the `data-slot` that
`Form` and `FormActions` write.
