+++ apps/web/app/projects/[projectId]/personas/page.tsx
@@ -15,7 +15,8 @@ import {
 } from "../../../../lib/personas.ts";
 import { projectLanding, projectPath } from "../../../../lib/project-context.ts";
 import { canAuthor } from "../../../../lib/roles.ts";
-import { Choice, Toolbar } from "../../../../ui/controls.tsx";
+import { Choice } from "../../../../ui/controls.tsx";
+import { Toolbar } from "../../../../ui/section.tsx";
 import { DataTable, type Column } from "../../../../ui/data-table.tsx";
 import { Empty, Failure, Loading, NotFound } from "../../../../ui/page-state.tsx";
 import { useProjectRead } from "../../../../ui/resource.ts";
