+++ apps/web/app/projects/[projectId]/runs/[runId]/page.tsx
@@ -22,7 +22,7 @@ import { simulationRerunPath } from "../../../../../lib/simulations.ts";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { cn } from "@/lib/utils";
-import { Actions, Section } from "../../../../../ui/controls.tsx";
+import { Actions, Section } from "../../../../../ui/section.tsx";
 import { Field, Refused } from "../../../../../ui/form.tsx";
 import { DataTable, type Column } from "../../../../../ui/data-table.tsx";
 import { Dialog } from "../../../../../ui/dialog.tsx";
