+++ apps/web/app/projects/[projectId]/runs/page.tsx
@@ -26,7 +26,8 @@ import {
 } from "../../../../lib/runs.ts";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Actions, Select, Toolbar } from "../../../../ui/controls.tsx";
+import { Actions, Toolbar } from "../../../../ui/section.tsx";
+import { Select } from "../../../../ui/controls.tsx";
 import { Refused } from "../../../../ui/form.tsx";
 import { DataTable, type Column } from "../../../../ui/data-table.tsx";
 import { Empty, Failure, Loading, NotFound } from "../../../../ui/page-state.tsx";
