+++ apps/web/ui/controls.tsx
@@ -522,74 +522,3 @@ export function Badge({
   );
 }
 
-/**
- * A labelled group of facts about one thing — what a detail page is mostly
- * made of. A definition list because that is what it is, so a screen reader
- * reads each fact with the name of the fact.
- */
-export function Facts({
-  facts,
-}: {
-  readonly facts: readonly {
-    readonly label: string;
-    readonly value: ReactNode;
-  }[];
-}) {
-  return (
-    <dl className={styles.facts}>
-      {facts.map((fact) => (
-        <div className={styles.fact} key={fact.label}>
-          <dt>{fact.label}</dt>
-          <dd>{fact.value}</dd>
-        </div>
-      ))}
-    </dl>
-  );
-}
-
-/**
- * A strip of controls above a list: a search box, the filters, and nothing that
- * belongs in the page header.
- *
- * It is here rather than in each list page so that every list in the product
- * puts its controls in the same place and at the same density. A page that
- * needs a fifth control puts it here beside the others rather than inventing a
- * second row.
- */
-export function Toolbar({ children }: { readonly children: ReactNode }) {
-  return <div className={styles.toolbar}>{children}</div>;
-}
-
-/** A group of controls that act on the thing the page is about. */
-export function Actions({ children }: { readonly children: ReactNode }) {
-  return <div className={styles.actions}>{children}</div>;
-}
-
-/** A titled block inside a page: connections, capabilities, history. */
-export function Section({
-  title,
-  lead,
-  action,
-  children,
-}: {
-  readonly title: string;
-  readonly lead?: ReactNode;
-  readonly action?: ReactNode;
-  readonly children: ReactNode;
-}) {
-  return (
-    <section className={styles.section}>
-      <header className={styles.sectionHead}>
-        <div>
-          <h2 className={styles.sectionTitle}>{title}</h2>
-          {lead === undefined ? null : (
-            <p className={styles.sectionLead}>{lead}</p>
-          )}
-        </div>
-        {action === undefined ? null : <div>{action}</div>}
-      </header>
-      {children}
-    </section>
-  );
-}
-
